requirements.txt
mysql-connector == 2.2.9
socket
datetime
pickle
threading